Move gdctx to bin/glusterd2/gdctx
