+++ bin/glusterd2/config.go
@@ -8,7 +8,7 @@ import (
 	"os"
 	"path"
 
-	"github.com/gluster/glusterd2/gdctx"
+	"github.com/gluster/glusterd2/bin/glusterd2/gdctx"
 	"github.com/gluster/glusterd2/store"
 	"github.com/gluster/glusterd2/volgen"
