+++ commands/peers/peer-rpc-svc.go
@@ -1,7 +1,7 @@
 package peercommands
 
 import (
-	"github.com/gluster/glusterd2/gdctx"
+	"github.com/gluster/glusterd2/bin/glusterd2/gdctx"
 	"github.com/gluster/glusterd2/peer"
 	"github.com/gluster/glusterd2/servers/peerrpc"
 	"github.com/gluster/glusterd2/store"
