+++ daemon/store.go
@@ -6,7 +6,7 @@ import (
 	"errors"
 	"path"
 
-	"github.com/gluster/glusterd2/gdctx"
+	"github.com/gluster/glusterd2/bin/glusterd2/gdctx"
 	"github.com/gluster/glusterd2/store"
 
 	"github.com/coreos/etcd/clientv3"
