+++ store/liveness.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"time"
 
-	"github.com/gluster/glusterd2/gdctx"
+	"github.com/gluster/glusterd2/bin/glusterd2/gdctx"
 
 	"github.com/coreos/etcd/clientv3"
 	"github.com/pborman/uuid"
