+++ transaction/lock.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 	"time"
 
-	"github.com/gluster/glusterd2/gdctx"
+	"github.com/gluster/glusterd2/bin/glusterd2/gdctx"
 	"github.com/gluster/glusterd2/store"
 
 	"github.com/coreos/etcd/clientv3/concurrency"
